ISIS-1670: changes ScalarPanelAbstract2 to use onInitialize to onBefore
NB: work-in-progress, this doesn't correctly work because doesn't install all the callbacks as required.  For example, in kitchensink, if bring up a ReferenceObject, and change "Some Other Object Optional With Choices" (as an editable property), first time works, but second time the callback to bring up the prompt is not installed.